Refactor patch to make the replication works
