+++ lib/octopus/model.rb
@@ -208,23 +208,3 @@ If you are trying to scope everything to a specific shard, use Octopus.using ins
 end
 
 ActiveRecord::Base.extend(Octopus::Model)
-
-module SampleFoooBar
-  attr_accessor :current_shard
-
-  private
-
-  def hash_rows
-    if current_shard.blank?
-      super
-    else
-      foo = super
-      foo.each { |f| f.merge!('current_shard' => current_shard) }
-      foo
-    end
-  end
-end
-
-class ActiveRecord::Result
-  prepend SampleFoooBar
-end
